# EasyAspCoreReactTemplate/ClientApp/src/api/models/UserInfoDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserInfoDto = {
  isAuthenticated?: boolean;
  userName?: string | null;
  email?: string | null;
};